restored c implementation and moved new one to c2
