+++ c/main.c
@@ -1,21 +1,15 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <stdint.h>
-#include <sys/mman.h>
-#include <sys/stat.h>
-#include <fcntl.h>
-#include <sys/types.h>
-#include <assert.h>
-
 
 
 int main( int argc, char **argv )
 {
   FILE *fin1;
   FILE *fin2;
-
-  char line[1024] = {0};
+  
+  char line1[1024];
+  char line2[1024];
 
   if( argc < 3 )
     return -1;
@@ -26,46 +20,27 @@ int main( int argc, char **argv )
   if( ( fin1 == NULL ) || (fin2 == NULL) )
     return -1;
 
-  struct stat st;
-  stat(argv[1], &st);
-  size_t finSize1 =  st.st_size;
-  stat(argv[2], &st);
-  size_t finSize2 =  st.st_size;
-
-  int fileNum1 = fileno(fin1);
-  int fileNum2 = fileno(fin2);
-  char* mmappedData1 = mmap(NULL, finSize1, PROT_READ, MAP_PRIVATE | MAP_POPULATE, fileNum1, 0);
-  assert(mmappedData1 != NULL);
-  char* mmappedData2 = mmap(NULL, finSize2, PROT_READ, MAP_PRIVATE | MAP_POPULATE, fileNum2, 0);
-  assert(mmappedData2 != NULL);
-
-  char * startOne = mmappedData1;
-  char * endOne = strchr(mmappedData1,'\n');
-  while (endOne!=NULL)
+  while( !feof(fin1) )
   {
-    int lengthOne = endOne-startOne;
-    memcpy(line, startOne, lengthOne);
-    char * startTwo = mmappedData2;
-    char * endTwo = strchr(mmappedData2,'\n');
-    while (endTwo!=NULL)
+    if( !fgets( line1, sizeof(line1), fin1 ) )
+      break;
+      
+    strtok( line1, "\n");
+
+    while( !feof(fin2) )
     {
-      int lengthTwo = endTwo-startTwo+1;
-      memcpy(line + lengthOne, startTwo, lengthTwo);
-      fwrite(line, 1 , lengthOne + lengthTwo, stdout);
-      startTwo = endTwo+1;
-      endTwo=strchr(endTwo+1,'\n');
+      if( !fgets( line2, sizeof(line2), fin2 ) )
+        break;
+    
+      fprintf( stdout, "%s%s", line1, line2 );
     }
-    startOne = endOne+1;
-    endOne=strchr(endOne+1,'\n');
+    
+    rewind( fin2 );
   }
-
-  int rc = munmap(mmappedData1, finSize1);
-  assert(rc == 0);
-  rc = munmap(mmappedData2, finSize2);
-  assert(rc == 0);
-
+  
+  
   fclose( fin1 );
   fclose( fin2 );
 
-  exit(0);
+  exit(0);  
 }
