+++ c2/.gitignore
@@ -1,0 +1,5 @@
+*
+*.*
+!.gitignore
+!main.c
+!Makefile
